Split out two separate Task Data classes, one for the index page and one for show
